# specs/004-identity/contracts/api.yaml
openapi: 3.0.3
info:
  title: RAGGit Workstation API (Single-Tenant)
  version: 1.3.0
  description: >
    Single-tenant on-prem RAG on AI Workstation (LanceDB local path +
    Ollama/LLamaSharp/ONNX). Desktop is thin client over LAN (no WAN at query
    time per FR-004/Constitution IV). 1.3.0 (004-identity, MINOR/ADDITIVE per
    FR-013): local per-person accounts — POST /api/auth/login and
    POST /api/auth/refresh issue workstation-signed JWT bearer tokens (8h);
    /api/users CRUD (Admin-only) manages the people directory; GET /api/auth/me
    gains ADDITIVE optional fields (displayName, username, sub) — 1.2.0
    required fields unchanged, clients MUST ignore unknown fields; Bearer
    security scheme added alongside retained ApiKey (machine/bootstrap per
    FR-012); credentials and session tokens traverse the LAN over HTTPS only
    (FR-008); 429 added for sign-in lockout (FR-009, 5 attempts / 15 min);
    Documents.createdBy and Queries attribution carry the stable person id for
    Local sessions while legacy "admin"/"employee" values remain valid
    (FR-007). No breaking changes: every 1.2.0 path, schema, and required
    field is preserved. 1.2.0 added xlsx MIME + rejection ordering; 1.1.0
    added GET /api/auth/me (extensible envelope) and pdf/docx 400 hardening.
servers:
  - url: https://ai-workstation.local:5001
    description: AI Workstation LAN — HTTPS required for credential and token traffic (FR-008)
  - url: http://localhost:5001
...
security:
  - ApiKey: []
  - Bearer: []
paths:
  /api/auth/login:
    post:
      summary: Sign in with local per-person credentials (004)
      description: >
        HTTPS only. Verifies the password against the PBKDF2-SHA256 (310000
        iterations, salted, non-recoverable) hash in the Users table (FR-002).
        Unknown username and wrong password return the IDENTICAL generic 401 —
        no user enumeration (US2.3). Five consecutive failures lock the account
        for 15 minutes: further attempts return 429 with Retry-After (FR-009).
        Inactive accounts receive the generic 401 (FR-010). Success issues an
        8-hour workstation-signed JWT (HS256) identifying the person, role, and
        display name (FR-003). Identity resolution is entirely local — no cloud
        IdP contact at any time; if the identity store is unavailable the API
        fails fast with 503, never a hang and never a remote fallback (FR-011).
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated — session credential issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials or inactive account — single generic error (no enumeration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Account locked — 5 failed sign-ins within 15 minutes (FR-009); retry after LockoutUntil
          headers:
            Retry-After:
              schema: { type: integer, example: 900 }
              description: Seconds until the lockout expires
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: identity store unavailable — fail fast, offline-consistent (FR-011)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/refresh:
    post:
      summary: Re-issue session credential from a still-valid token (004, optional — P3)
      description: >
        Presents a current unexpired Bearer token and receives a fresh 8h
        token. The client uses this opportunistically before expiry; when the
        token has already expired the client MUST re-prompt sign-in — never a
        silent fallback to a shared API key for person actions (US4.2).
      operationId: refreshToken
      security:
        - Bearer: []
      responses:
        '200':
          description: New session credential issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      summary: Authenticated identity and role (for thin client UI gating)
      description: >
        Returns the caller's identity envelope. Extensible — client MUST ignore
        unknown fields (Q2 forward-compat). Role determines Admin vs Employee
        UI (Upload/Delete gating). Added 1.1.0 MINOR. 1.3.0 ADDITIVE —
        displayName, username, and sub are returned for Local (per-person JWT)
        sessions; ApiKey (machine/bootstrap) sessions keep returning the 1.2.0
        shape unchanged (FR-006, FR-013). sub is the stable person id used for
        attribution (FR-007).
      operationId: getAuthMe
      security:
        - ApiKey: []
        - Bearer: []
      responses:
        '200':
          description: Authenticated identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMe'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: >
        Single-tenant machine/bootstrap key (FR-012). Retained for machine
        integration and initial bring-up; NOT used for per-person attribution
        when a Local session exists (FR-007). Windows AD remains a documented
        future provider.
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Workstation-signed HS256 JWT issued by POST /api/auth/login (004).
        8-hour lifetime; identifies a person (sub = stable person id), role,
        and display name. HTTPS-only transport (FR-008). Deactivated or
        locked-out accounts are refused per request (FR-010).
  schemas:
    LoginRequest:
      type: object
      properties:
        username: { type: string, minLength: 3, maxLength: 64, example: ada }
        password: { type: string, format: password, minLength: 1 }
      required: [username, password]
    TokenResponse:
      type: object
      properties:
        access_token: { type: string, description: JWT (HS256) — present only over HTTPS }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, example: 28800, description: Seconds — default 8h (Auth:TokenLifetimeHours) }
      required: [access_token, token_type, expires_in]
...
    AuthMe:
      type: object
      description: >
        Extensible identity envelope (Q2). Client ignores unknown fields.
        1.3.0 ADDITIVE — displayName, username, sub are present for Local
        (per-person) sessions; ApiKey sessions return the 1.2.0 shape
        (identityType + role only).
      properties:
        identityType: { type: string, enum: [ApiKey, Local], example: Local, description: "'Local' = per-person JWT session (004); 'ApiKey' = machine/bootstrap (002). Future providers (e.g. AD) are additive." }
        role: { type: string, enum: [Admin, Employee] }
        displayName: { type: string, description: ADDED 1.3.0 (optional) — Local sessions only }
        username: { type: string, description: ADDED 1.3.0 (optional) — Local sessions only }
        sub: { type: string, format: uuid, description: ADDED 1.3.0 (optional) — stable person id; attribution key }
      required: [identityType, role]
      additionalProperties: true
...
    Error:
      type: object
      properties:
        error: { type: string, example: unauthorized }
  responses:
...
    Unauthorized:
      description: Unauthorized — missing/invalid/expired credential, or inactive account (generic)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...